Fix testing define_mounted_helper
Check for define_mounted_helper was added at 28cf7720 and FakeEngine was
created to test mounting an object with routes method which is not a
RouteSet. In e6c602da9 FakeEngine was changed which allowed to remove
define_mounted_helper check without breaking tests.

In order to avoid such regressions here I create a different class for
testing mounting inside a resource.